requirements.txt
asciimatics==1.11.0
future==0.18.2
hashids==1.2.0
Pillow==7.1.2
pkg-resources==0.0.0
pyfiglet==0.8.post1
python-gnupg==0.4.6
wcwidth==0.1.9
